interfaced all the controllers
